+++ app/src/main/res/layout/activity_filepicker.xml
@@ -3,12 +3,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="is.xyz.mpv.FilePickerActivity">
+    tools:context="is.xyz.filepicker.FilePickerActivity">
 
     <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/fragment_container_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/fragment_container_view"
         tools:layout="@layout/nnf_fragment_filepicker" />
 
     <TextView
