+++ app/src/main/res/values-tr/strings.xml
@@ -91,9 +91,6 @@
     <string name="pref_auto_rotation_title">Ekran yönü</string>
     <string name="pref_auto_rotation_summary">Yatay veya dikey videoları mpv\'nin hangi yönde oynatacağına karar verin</string>
 
-    <string name="pref_display_media_title_title">Medya başlığını göster (videolar)</string>
-    <string name="pref_display_media_title_summary">Videoların medya başlığı veya dosya adı denetimlerin üzerinde görüntülenir. Ses dosyalarının üst verileri her zaman gösterilir.</string>
-
     <string name="pref_bottom_controls_title">Denetimleri ekranın altında göster</string>
 
     <string name="pref_no_ui_pause_title">Kullanıcı arayüzü açılır pencereleri sırasında oynatmaya devam et</string>
