+++ app/src/main/res/values-uk/strings.xml
@@ -98,9 +98,6 @@
     <string name="pref_auto_rotation_title">Орієнтація екрана</string>
     <string name="pref_auto_rotation_summary">Визначає, в якій орієнтації mpv буде відтворювати ландшафтні чи портретні відео.</string>
 
-    <string name="pref_display_media_title_title">Показувати заголовки (для відео)</string>
-    <string name="pref_display_media_title_summary">Показувати заголовок чи назву відеофайлу над елементами керування. Для аудіофайлів завжди будуть виводитись їхні метадані.</string>
-
     <string name="pref_bottom_controls_title">Показувати елементи керування знизу екрана</string>
 
     <string name="pref_no_ui_pause_title">Продовжувати відтворення, коли є додаткові вікна</string>
